acabando back

+++ code/lds3/src/main/java/br/com/pucminas/lds3/model/Professor.java
@@ -9,9 +9,9 @@ public class Professor extends User {
     private String departamento;
 
     @Column(nullable = false, length = 100)
-    private short saldo;
+    private double saldo;
 
-    public Professor(String nome, String email, String password, String departamento, short saldo) {
+    public Professor(String nome, String email, String password, String departamento, double saldo) {
         super(nome, email, password);
         this.departamento = departamento;
         this.saldo = saldo;
@@ -25,11 +25,11 @@ public void setDepartamento(String departamento) {
         this.departamento = departamento;
     }
 
-public short getSaldo() {
+public double getSaldo() {
         return saldo;
     }
 
-public void setSaldo(short saldo) {
+public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
